Project structure with list of todo sorted
